fix(striker-ui): revise inner and expandable panel types
