+++ striker-ui/components/Panels/InnerPanel.tsx
@@ -1,28 +1,26 @@
-import { FC } from 'react';
-import { Box as MUIBox, BoxProps as MUIBoxProps } from '@mui/material';
+import { FC, useMemo } from 'react';
+import { Box as MUIBox, SxProps, Theme } from '@mui/material';
 
 import { BORDER_RADIUS, DIVIDER } from '../../lib/consts/DEFAULT_THEME';
 
-type InnerPanelProps = MUIBoxProps;
+const InnerPanel: FC<InnerPanelProps> = ({ sx, ...muiBoxRestProps }) => {
+  const combinedSx = useMemo<SxProps<Theme>>(
+    () => ({
+      borderWidth: '1px',
+      borderRadius: BORDER_RADIUS,
+      borderStyle: 'solid',
+      borderColor: DIVIDER,
+      marginTop: '1.4em',
+      marginBottom: '1.4em',
+      paddingBottom: 0,
+      position: 'relative',
 
-const InnerPanel: FC<InnerPanelProps> = ({ sx, ...muiBoxRestProps }) => (
-  <MUIBox
-    {...{
-      sx: {
-        borderWidth: '1px',
-        borderRadius: BORDER_RADIUS,
-        borderStyle: 'solid',
-        borderColor: DIVIDER,
-        marginTop: '1.4em',
-        marginBottom: '1.4em',
-        paddingBottom: 0,
-        position: 'relative',
+      ...sx,
+    }),
+    [sx],
+  );
 
-        ...sx,
-      },
-      ...muiBoxRestProps,
-    }}
-  />
-);
+  return <MUIBox {...muiBoxRestProps} sx={combinedSx} />;
+};
 
 export default InnerPanel;
