+++ striker-ui/types/ExpandablePanel.d.ts
@@ -1,0 +1,10 @@
+type ExpandablePanelOptionalProps = {
+  expandInitially?: boolean;
+  loading?: boolean;
+  panelProps?: InnerPanelProps;
+  showHeaderSpinner?: boolean;
+};
+
+type ExpandablePanelProps = ExpandablePanelOptionalProps & {
+  header: import('react').ReactNode;
+};
