+++ striker-ui/types/InnerPanel.d.ts
@@ -1,0 +1,1 @@
+type InnerPanelProps = import('@mui/material').BoxProps;
